remove
`<script src="{{"js/html5shiv-printshiv.min.js" | relURL}}{{ if not .Site.Params.disableAssetsBusting }}?{{ now.Unix }}{{ end }}"></script>`
This was only ever needed for IE8 (which is not supported with this theme since it uses a newer version of jQuery). Furthermore it should be loaded in the head anyway. IE9 supports HTML5 tags and does not need this polyfill at all.